V1.8.64 - Updates
- Cleaned up Meade comments
- Fixed a bug that caused issues when you turned off microstepping.
- Fixed a bug that prevented DEC from using floating point steps per degree.

+++ src/Mount.hpp
@@ -317,7 +317,7 @@ private:
 private:
   LcdMenu* _lcdMenu;
   float _stepsPerRADegree;
-  int _stepsPerDECDegree;
+  float _stepsPerDECDegree;
   int _maxRASpeed;
   int _maxDECSpeed;
   int _maxRAAcceleration;
